+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - fileType to history data
 - change page title when editing a file
 - jwt library update
+- fileType parameter used in callback instead of extension from url (DocumentServer 7.0 required)
 
 ## 8.1.0
 ## Changed